fixed for proxy bug

+++ ecmascript/compiler/call_signature.cpp
@@ -520,7 +520,7 @@ DEF_CALL_SIGNATURE(CallRuntimeWithArgv)
     std::array<VariableType, 4> params = { /* 4 : 4 input parameters */
         VariableType::NATIVE_POINTER(), // glue
         VariableType::INT64(),   // runtimeId
-        VariableType::INT32(),   // argc
+        VariableType::INT64(),   // argc
         VariableType::NATIVE_POINTER(), // argv
     };
     callSign->SetVariadicArgs(false);
@@ -537,8 +537,8 @@ DEF_CALL_SIGNATURE(OptimizedCallOptimized)
     *callSign = runtimeCallTrampoline;
     std::array<VariableType, 4> params = { /* 4 : 4 input parameters */
         VariableType::NATIVE_POINTER(),
-        VariableType::INT32(),
-        VariableType::INT32(),
+        VariableType::INT64(),
+        VariableType::INT64(),
         VariableType::NATIVE_POINTER(),
     };
     callSign->SetVariadicArgs(true);
@@ -556,7 +556,7 @@ DEF_CALL_SIGNATURE(JSCall)
     std::array<VariableType, 6> params = { // 6 : 6 input parameters
         VariableType::NATIVE_POINTER(),     // glue
         VariableType::JS_ANY(),      // lexenv
-        VariableType::INT32(),       // actual argC
+        VariableType::INT64(),       // actual argC
         VariableType::JS_ANY(),      // call target
         VariableType::JS_ANY(),      // new target
         VariableType::JS_ANY(),      // thisobj
@@ -575,7 +575,7 @@ DEF_CALL_SIGNATURE(JSProxyCallInternalWithArgV)
     *callSign = jSProxyCallInternalWithArgV;
     std::array<VariableType, 4> params = { // 4 : 4 input parameters
         VariableType::NATIVE_POINTER(),     // glue
-        VariableType::INT32(),       // actual argC
+        VariableType::INT64(),       // actual argC
         VariableType::JS_ANY(),      // call target
         VariableType::NATIVE_POINTER(),    // argv
     };
@@ -594,8 +594,8 @@ DEF_CALL_SIGNATURE(JSFunctionEntry)
     std::array<VariableType, 6> params = {  // 6 : 6 input parameters
         VariableType::NATIVE_POINTER(),     // glue
         VariableType::NATIVE_POINTER(),     // prev fp
-        VariableType::INT32(),              // expectedNumArgs
-        VariableType::INT32(),              // actualNumArgs
+        VariableType::INT64(),              // expectedNumArgs
+        VariableType::INT64(),              // actualNumArgs
         VariableType::NATIVE_POINTER(),     // argv
         VariableType::NATIVE_POINTER(),     // codeAddr
     };
@@ -800,7 +800,7 @@ DEF_CALL_SIGNATURE(JSCallWithArgV)
     // 6 : 6 input parameters
     std::array<VariableType, 6> params = {
         VariableType::NATIVE_POINTER(),   // glue
-        VariableType::INT32(),            // actualNumArgs
+        VariableType::INT64(),            // actualNumArgs
         VariableType::JS_ANY(),           // jsfunc
         VariableType::JS_ANY(),           // newTarget
         VariableType::JS_ANY(),           // this
@@ -1024,7 +1024,7 @@ DEF_CALL_SIGNATURE(JsProxyCallInternal)
     *callSign = proxyCallInternal;
     std::array<VariableType, 4> params = { // 4 : 4 input parameters
         VariableType::NATIVE_POINTER(),    // glue
-        VariableType::INT32(),      // actual argC
+        VariableType::INT64(),      // actual argC
         VariableType::JS_POINTER(), // callTarget
         VariableType::NATIVE_POINTER(),    // argv
     };
